pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
version = "0.9.0dev"
name = "dtcc-viewer"
description = "DTCC Viewer"
readme = "README.md"
authors = [
  { name = "Jens Olsson", email = "jens.olsson@chalmersindustriteknik.se" },
  { name = "Dag Wästberg", email = "dwastberg@gmail.com" },
]
requires-python = ">=3.10,<3.13"
classifiers = [
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: C++",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "jupyter",
  "ipywidgets",
  "matplotlib",
  "colormap",
  "numpy",
  "folium >= 0.14.0",
  "pyglet",
  "trimesh",
  "pyopengl",
  "pyglfw",
  "pyrr",
  "pillow",
  "glfw",
  "imgui",
  "tqdm>=4.66.1",
  "dtcc-core@git+https://github.com/dtcc-platform/dtcc-core.git@develop", 
]

[project.optional-dependencies]
test = ["pytest"]
dev = ["easydev","black"]
